requirements.txt
selenium>=4.13.0
viam-sdk>=0.2.18
